Linear Regression
Implement linear regression with one variable to predict proﬁts for a food truck. The CEO of a restaurant franchise and are considering diﬀerent cities for opening a new
outlet. The chain already has trucks in various cities and you have data for proﬁts and populations from the cities.